requirements.txt
argon2-cffi>=19.1.0
cffi>=1.12.3
dj-database-url>=0.5.0
Django>=2.2.2
django-authtools>=1.6.0
djangorestframework>=3.9.4
psycopg2-binary>=2.8.2
pycparser>=2.19
python-decouple>=3.1
pytz>=2019.1
six>=1.12.0
sqlparse>=0.3.0
